+++ schnorr_fun/benches/bench_schnorr.rs
@@ -8,7 +8,7 @@ use sha2::Sha256;
 const MESSAGE: &'static [u8; 32] = b"hello world you are beautiful!!!";
 
 lazy_static::lazy_static! {
-    static ref SK: Scalar<Secret,NonZero> = Scalar::from_bytes_mod_order(*b"xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx").mark::<NonZero>().unwrap();
+    static ref SK: Scalar<Secret,NonZero> = Scalar::from_bytes_mod_order(*b"xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx").non_zero().unwrap();
     static ref schnorr: Schnorr<Sha256, Deterministic<Sha256>> = Schnorr::new(Deterministic::default());
 }
 
@@ -47,7 +47,7 @@ fn verify_schnorr(c: &mut Criterion) {
         });
 
         {
-            let sig = sig.clone().mark::<Secret>();
+            let sig = sig.clone().set_secrecy::<Secret>();
             group.bench_function("fun::schnorr_verify_ct", |b| {
                 b.iter(|| schnorr.verify(&verification_key, message, &sig))
             });