+++ secp256kfun/src/marker/point_type.rs
@@ -2,20 +2,12 @@
 ///
 /// There are several different point types.
 /// - [`Normal`]: A point represented internally with Affine coordinates, with x and y coordinates (if it's not zero). These can be directly serialized or hashed.
-/// - [`Jacobian`]: A non-normal represented internally in Jacobian coordinates. Usually the result of a point operation. Before being serialized or hashed, you have to normalize it.
-/// - [`BasePoint`]: A normal point that has pre-computed multiplication tables like [`G`].
+/// - [`NonNormal`]: A non-normalized represented internally as in three   coordinates. Usually the result of a point operation. Before being serialized or hashed, you have to normalize it.
+/// - [`BasePoint`]: A normal point that has (or may have in the future) pre-computed multiplication tables like [`G`].
 /// - [`EvenY`]: A normal point whose y-coordinate is known to be _even_ at compile time.
 ///
-/// To normalize a `Point<Jacobian>` you mark it as [`Normal`]:
-/// ```
-/// use secp256kfun::{g, marker::*, Scalar, G};
-/// let scalar = Scalar::random(&mut rand::thread_rng());
-/// let jacobian_point = g!(scalar * G);
-/// let normal_point = jacobian_point.mark::<Normal>();
-/// let bytes = normal_point.to_bytes(); // we can now serialize it
-/// ```
-/// [`G`]: crate::G
 /// [`Point<T,S,Z>`]: crate::Point
+/// [`G`]: crate::G
 pub trait PointType: Sized + Clone + Copy + 'static {
     /// The point type returned from the negation of a point of this type.
     type NegationType: Default;
@@ -26,15 +18,29 @@ pub trait PointType: Sized + Clone + Copy + 'static {
 #[derive(Default, Debug, Clone, Copy)]
 #[cfg_attr(feautre = "serde", derive(serde::Serialize, serde::Deserialize))]
 pub struct Normal;
+/// A Non-normalized Point. Usually, represented as three field elements three field elements:
+/// `x`,`y` and `z` rather than just two in a [`Normal`] point.
+///
+/// In general it's most efficient to normalize `NonNormal` points into `Normal` points, as late as
+/// possible. To normalize a `NonNormal` point call `normalize` on the point.
+///
+/// ```
+/// use secp256kfun::{g, marker::*, Scalar, G};
+/// let scalar = Scalar::random(&mut rand::thread_rng());
+/// let non_normal_point = g!(scalar * G);
+/// let normal_point = non_normal_point.normalize();
+/// let bytes = normal_point.to_bytes(); // we can now serialize it
+/// ```
+///
+/// [`normalize`]: crate::Point::normalize
+
 #[derive(Default, Debug, Clone, Copy)]
-/// A Non-normalized Point. `Jacobian` points are represented internally as
-/// three field elements: `x`,`y` and `z`. Most point operations return
-/// `Jacobian` points.
-///
-/// In general it's most efficient to normalize `Jacobian` points into [`Normal`] points, as late as possible.
-///
-/// Note that the underlying arithmetic backend may not actually use Jacobian coordinates - it may use Projective coordinates instead.
-pub struct Jacobian;
+pub struct NonNormal;
+
+/// Backwards compatibility type alias.
+#[deprecated(note = "use NonNormal instead")]
+pub type Jacobian = NonNormal;
+
 /// A [`Normal`] point whose `y` coordinate is known to be even.
 #[derive(Default, Debug, Clone, Copy)]
 #[cfg_attr(feautre = "serde", derive(serde::Serialize, serde::Deserialize))]
@@ -58,7 +64,7 @@ impl Normalized for EvenY {}
 impl Normalized for Normal {}
 impl Normalized for BasePoint {}
 
-impl NotBasePoint for Jacobian {}
+impl NotBasePoint for NonNormal {}
 impl NotBasePoint for EvenY {}
 impl NotBasePoint for Normal {}
 
@@ -66,26 +72,6 @@ impl<N: Normalized> PointType for N {
     type NegationType = Normal;
 }
 
-impl PointType for Jacobian {
-    type NegationType = Jacobian;
-}
-
-mod change_marks {
-    use crate::{marker::*, Point};
-
-    impl<T, S, Z> ChangeMark<Point<T, S, Z>> for Normal {
-        type Out = Point<Normal, S, Z>;
-
-        fn change_mark(point: Point<T, S, Z>) -> Self::Out {
-            crate::op::point_normalize(point)
-        }
-    }
-
-    impl<Z, S, T> ChangeMark<Point<T, S, Z>> for Jacobian {
-        type Out = Point<Jacobian, S, Z>;
-
-        fn change_mark(point: Point<T, S, Z>) -> Self::Out {
-            Point::from_inner(point.0, Jacobian)
-        }
-    }
+impl PointType for NonNormal {
+    type NegationType = NonNormal;
 }